fix(openapi): improve parameter processing
The spec https://spec.openapis.org/oas/latest.html#fixed-fields-7 says
the name + location must be unique, so the same name *IS* allowed in a
different location. This would cause us to have the parameter to
overwrite the other, wich we tried to solve by prepending "req-",
which is far from good.

Now we build a map-of-maps with the parameter name, followed by
location and then the parameter+name structure, then we iterate over
it.

The iteration is also shared between parameter declaration and value
usage, avoiding any inconsistencies.

More than that, if the parameter existed in both path and operation,
it was not being properly overridden.

Last but not least, buildRequestFromParams() was fixed:

- do not replace path parameters in the server url

- do not change the server url once it's created (not needed)

- close the req body on failures, otherwise file descriptors may be
  left open (ie: when uploading)