feat(tf): add action resources to TF
Now terraform can handle sub resources like instance backup, snapshots
and can also express relationship between resources as other resources.

This is done using the links in each operation, for example, for
attaching a volume to a virtual machine we need to defined the oposite
operation as a `delete` link (detach in this case) and we must provide a
read operation with a termination check to validate if the relationship
is valid when reading the current infrastructure state.

Operations that don't modify resource states like "get" are not allowed
to become action resouces.