feat(scripts): stats for unused schema components
Components may be of multiple types. They can be defined inside the
`components` key and locally referenced from most places of the spec
with `$ref: "#/components/<type>/<name>"`. For now, reusing the
"generic" get function to resolve nodes in the traversal to populate
a map when a ref is found.