feat: customizable config dir (#1469)
* fix: properly declare python hooks and their deps

On strict environments (like nix shells) the hooks run on sandboxed
environments, so any dependency they require should be declared
as such.

* feat: allows customization of the config file dir

The current profile management and switching implementation is
stateful, so users working with multiple accounts at the same
time, with different profiles on different sessions or console
windows can get their enviromments mixed up.

This implementation allows the configuration directory to be
defined by an environment variable, so users with that specific
need can have with fully isolated profile management per project or
customer.

Note that the usage is completely optional: when the MGC_CONFIG_DIR
environment variable is not set, the regular configuration path
resolution kicks in.